some columns added in entiy

+++ CoreEntities/Transaction.cs
@@ -10,7 +10,7 @@ namespace SplitExpenses.Entities
         public string ParticipantName { get; set; }
         public string GroupName { get; set; }
         public bool IsActive { get; set; }
-        public int Amount { get; set; }
+        public decimal Amount { get; set; }
         public string Remarks { get; set; }
         public string ExtraInfo { get; set; }
         public string ExtraInfo1 { get; set; }
